클래스 등록 postman으로 확인 완료

+++ src/main/java/com/knocksea/see/edu/repository/EduRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 
-public interface EduRepository extends JpaRepository<Edu,Integer> {
+public interface EduRepository extends JpaRepository<Edu,Long> {
 }
